feat: create shared layout for login and register pages
